# src/main/resources/openapi.yaml
# open api speicification document - header with version
# document should be named openapi.yaml or openapi.json
openapi: "3.0.4"
info:
  title: Beverage Service
  version: "1.0"
  description: >
    Beverage Service API — catalog and ordering endpoints.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:9999/v1

components:
  schemas:

    bottleDTO:
      type: object
      required:
        - name
        - volume
        - price
        - inStock
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        volume:
          type: number
          format: double
        price:
          type: number
          format: double
        inStock:
          type: integer
          format: int32
        isAlcoholic:
          type: boolean
        volumePercent:
          type: number
          format: double
        supplier:
          type: string
        href:
          type: string
          format: uri
      xml:
        name: bottle

    crateDTO:
      type: object
      required:
        - bottle
        - noOfBottles
        - price
      properties:
        id:
          type: integer
          format: int32
        bottle:
          $ref: "#/components/schemas/bottleDTO"
        noOfBottles:
          type: integer
          format: int32
        price:
          type: number
          format: double
        inStock:
          type: integer
          format: int32
        href:
          type: string
          format: uri
      xml:
        name: crate

    beverageDTO:
      description: A beverage payload — either a bottle or a crate.
      oneOf:
        - $ref: "#/components/schemas/bottleDTO"
        - $ref: "#/components/schemas/crateDTO"
      discriminator:
        propertyName: "type"
      xml:
        name: beverage

    beverageDTOList:
      type: array
      items:
        $ref: "#/components/schemas/beverageDTO"

    orderItemDTO:
      type: object
      required:
        - number
        - beverageDTO
        - quantity
      properties:
        number:
          type: integer
          format: int32
        beverageDTO:
          $ref: "#/components/schemas/beverageDTO"
        quantity:
          type: integer
          format: int32
      xml:
        name: orderItem

    orderDTO:
      type: object
      required:
        - orderItemDTOS
        - price
        - status
      properties:
        orderId:
          type: integer
          format: int32
        orderItemDTOS:
          type: array
          items:
            $ref: "#/components/schemas/orderItemDTO"
        price:
          type: number
          format: double
        status:
          type: string
          description: Order status (e.g. SUBMITTED, PROCESSED, CANCELLED)
        href:
          type: string
          format: uri
      xml:
        name: order

    pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
        noOfPages:
          type: integer
          format: int32
        first:
          type: string
          format: uri
        previous:
          type: string
          format: uri
        next:
          type: string
          format: uri
        last:
          type: string
          format: uri

    paginatedBeverage:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/pagination"
        dtos:
          type: array
          items:
            $ref: "#/components/schemas/beverageDTO"
        href:
          type: string
          format: uri

    paginatedOrder:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/pagination"
        dtos:
          type: array
          items:
            $ref: "#/components/schemas/orderDTO"
        href:
          type: string
          format: uri

    errorMessage:
      type: object
      required:
        - errorType
      properties:
        errorType:
          description: error category
          enum:
            - INVALID_PARAMETER
        message:
          type: string
          description: A detailed error description

paths:

  /beverage:
    get:
      summary: Get beverages (paginated)
      parameters:
        - name: pageNo
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageLimit
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "200":
          description: Paginated beverages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/paginatedBeverage"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Beverage Endpoints

  /beverage/search:
    get:
      summary: Get beverages between min and max price
      parameters:
        - name: min
          in: query
          schema:
            type: number
            format: double
            default: 0
        - name: max
          in: query
          schema:
            type: number
            format: double
            default: 10
      responses:
        "200":
          description: List of beverages within the given range
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/beverageDTO"
      tags:
        - Beverage Endpoints

  /order:
    get:
      summary: Get orders (paginated)
      parameters:
        - name: pageNo
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageLimit
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/paginatedOrder"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Order Endpoints

    post:
      summary: Submit an order (list of beverages)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/beverageDTOList"
      responses:
        "201":
          description: Order submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/paginatedOrder"
        "400":
          description: Message body was empty or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Order Endpoints

  /order/{orderId}:
    get:
      summary: Get an order by id
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderDTO"
        "404":
          description: Not found
      tags:
        - Order Endpoints

  /order/changeOrderStatus/{orderId}/{orderStatus}:
    get:
      summary: Change order status (customer)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: orderStatus
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderDTO"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Order Endpoints

  /order/updateOrder:
    put:
      summary: Update an existing order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orderDTO"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderDTO"
        "400":
          description: Can't update order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Order Endpoints

  /employee/changeOrderStatus/{orderId}/{orderStatus}:
    get:
      summary: Change order status (employee)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: orderStatus
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderDTO"
        "400":
          description: Can't update order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Employee Endpoints

  /employee/createNewBeverages:
    post:
      summary: Create new beverages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/beverageDTOList"
      responses:
        "201":
          description: Beverages created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/beverageDTOList"
        "400":
          description: Message body was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Employee Endpoints

  /employee/updateBeverage:
    put:
      summary: Update a beverage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/beverageDTO"
      responses:
        "200":
          description: Beverage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/beverageDTO"
        "400":
          description: Can't update beverage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
      tags:
        - Employee Endpoints